Check for 3d stacks and add get paths functionality
